Add continuous sampling via /dev/iio to ADC
This adds a bunch of patches to the adc patchset enabling continous
sampling mode. A large buffer of data can now be read via /dev/iio entries

The work is forward ported from the Arago tree.
The original author is Rachil Patna.

Some of his patches might seem missing from Arago but that is because

1) I incorporated *some* of his minor bugfixes in the same large main patch.
2) Some of the patches have already been incorporated in the patchset in mfd-next.

The rest of the small bug fixes by Russ Dill are kept in separate patches.

All in all. This brings all the functionality of the ADC driver in the Arago tree
to the BBB.