readme: add CROSS_COMPILE and ARCH flags into kernel make command
It can be a bit confusing to a new user when they first try to compile the kernel as per these instructions, as it will use the host environment to compile the kernel, which will fail. Add in example CROSS_COMPILE value, and add ARCH=arm to at least send them down the right path.